+++ lib/src/webview_options.dart
@@ -89,7 +89,7 @@ class InAppWebViewOptions
   ///Set to `true` to be able to listen at the [onDownloadStart] event. The default value is `false`.
   bool useOnDownloadStart;
 
-  ///Set to `true` to have all the browser's cache cleared before the new window is opened. The default value is `false`.
+  ///Set to `true` to have all the browser's cache cleared before the new WebView is opened. The default value is `false`.
   bool clearCache;
 
   ///Sets the user-agent for the WebView.